requirements.txt
minecraft-launcher-lib==8.0
pillow==12.0.0
psutil==7.1.2
PyQt6==6.10.0
screeninfo==0.8.1
requests==2.32.5
supabase==2.27.2